feat: Add react-helmet dependency and update Header component
Add the react-helmet dependency to the project by updating the package.json file. This will allow us to dynamically update the page title based on the current route. Also, update the Header component to import and use the Helmet component from react-helmet. This will enable us to set the page title for each page in the application, providing better SEO and user experience.